requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai-whisper>=20231117
language_tool_python>=2.7.1
